moved stuff to rust lib
