+++ rapl-rust-lib/Cargo.toml
@@ -8,3 +8,15 @@ crate-type = ["cdylib"]
 
 [dependencies]
 libc = "0.2"
+
+[dependencies.windows]
+version = "0.51"
+features = [
+    "Win32_Storage_FileSystem",
+    "Win32_Foundation",
+    "Win32_System_IO",
+    "Win32_Storage_FileSystem",
+    "Win32_Security",
+    "Win32_System_Threading",
+    "Win32_Media_DeviceManager",
+]
