+++ rapl-windows/Cargo.toml
@@ -8,15 +8,3 @@ edition = "2021"
 [dependencies]
 anyhow = "1"
 sysinfo = "0.29"
-
-[dependencies.windows]
-version = "0.51"
-features = [
-    "Win32_Storage_FileSystem",
-    "Win32_Foundation",
-    "Win32_System_IO",
-    "Win32_Storage_FileSystem",
-    "Win32_Security",
-    "Win32_System_Threading",
-    "Win32_Media_DeviceManager",
-]
